feat: allow profile picture update in winger mode
Extract AvatarPicker to a shared component and wire it into the winger
Me tab so wingers can update their avatar regardless of mode.
